25.04 Первый ревью
Упростили файл routes.py, перенесли функционал работы с файлами в file_manager
