+++ templates/files.html
@@ -4,23 +4,24 @@
     <title>{{ username }}'s Files</title>
 </head>
 <body>
-    <h1>{{ username }}'s Uploaded Files</h1>
+    <h1>{{ username }}'s Загруженные файлы</h1>
     <ul>
     {% for file in files %}
         <li>
             {{ file }}
-            <a href="{{ url_for('download_file', filename=file) }}">Download</a>
+            <button><a href="{{ url_for('download_file', filename=file) }}">Загрузить</a></button>
             <form action="{{ url_for('delete_file', filename=file) }}" method="post">
-                <button type="submit">Delete</button>
+                <button type="submit">Удалить</button>
             </form>
         </li>
     {% endfor %}
     </ul>
-    <h2>Upload new file</h2>
+    <h2>Загрузить новый файлы</h2>
     <form method="post" enctype="multipart/form-data">
         <input type="file" name="file">
-        <input type="submit" value="Upload">
+        <input type="submit" value="Загрузить">
     </form>
-    <a href="{{ url_for('logout') }}">Logout</a>
+    <button><a href="{{ url_for('logout') }}">Выйти</a></button>
+    <button><a href="Personal_Page">Личный кабинет</a></button>
 </body>
 </html>
