+++ templates/login.html
@@ -12,6 +12,7 @@
             <input type="password" name="password" placeholder="Пароль" required>
             <button type="submit">Войти</button>
         </form>
+        <button><a href="/register">Регистрация</a></button>
     </div>
 </body>
 </html>
