+++ templates/register.html
@@ -1,17 +1,18 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>Register</title>
+    <title>Регистрация</title>
 </head>
 <body>
-    <h1>Register</h1>
+    <h1>Регистрация</h1>
     {% if error %}
         <p style="color: red;">{{ error }}</p>
     {% endif %}
     <form method="post">
-        Username: <input type="text" name="username" required>
-        Password: <input type="password" name="password" required>
-        <input type="submit" value="Register">
+        Имя пользователя: <input type="text" name="username" required>
+        Пароль: <input type="password" name="password" required>
+        <input type="submit" value="Регистрация">
     </form>
+    <button><a href="/login">Логин</a></button>
 </body>
 </html>
