+++ templates/settings.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Настройки</title>
+</head>
+<body>
+  <h1>Настройки</h1>
+  <form>
+    <input type="submit" value="Сменить пароль">
+    <input type="submit" value="Сменить имя">
+  </form>
+</body>
+</html>
